Simplify & generalize stream read behavior at initial time
This commit removes the restriction that the initial condition stream
have the name 'input' and now allows any streams of direction input that
have input_interval='initial_only' to be read at the initial time.  This
allows an initial condition to be spread across multiple files.

It also prevents an obscure error that was causing streams of
input_interval='initial_only' to be read on the initial time of a
restart.  With this change, 'initial_only' truly means the initial time
of a simulation, regardless of restarting it or not.

This commit also adds some missing timer calls.