Modify reading of SLM mapping files to use mpas i/o instead of netcdf
This seems to have solved a strange issue of the SLM hanging when many
streams files were present.  The behavior was not entirely reproducible,
but initial testing seems to indicate that this change has solved it.
Note however, that this commit did not solve the problem until I added
the lines to close the files at the end of the routine.  It might have
been sufficient to keep the old netcdf routine but add lines for
closing the files when done.