Change definition of floating dynamic ice
This commit changes the deifnition of dynamic floating ice.  The old
method admitted edge cases where small thickness changes would result in
activation or deactivation of dynamic floating cells near grounding
lines that cause different forms of velocity mesh extension and very
large velocity differences after small (order meter) changes in
thickness.

The new method keeps track of the previous time step's dynamic mask.  If
a location was previously non-dynamic, it will remain that way until the
thickness criterion is reached.  If a location was previously dynamic,
it will remain that way until it thins away.

Note that this approach requires initializing and keeping track of the
previous cellMask, which required changes to Registry and the driver.
Also note that this has implications for restarts because a restart and
a cold start from a restart file will no longer give identical results.
Thus, it is advised to use restarts for branch runs, instead of making
the new runs.