Fix restarts, move code
This commits fixes restarts by adding cellMaskPrevious to the restart
stream.  In principle, it should be possible to avoid needing to include
it, but there are a lot of complications with the mask routine being
called many times during init that would need to be avoided on a restart to allow
the value of cellMask read from a restart value to be assigned to
cellMaskPrevious at the start of the fist real timestep.  Rather than
tackling that and potentially having a fragile solution, I took the
easier brute force approach of adding the additional restart field.

This commit also moves some of the new operations to the mask module to
avoid cluttering the main driver.