Stop doing extrapolation for Albany mesh for marine margins
This commit adjusts the mask of where Albany should be active to not
extend the mesh for marine margins.  Terrestrial margins are still
extended to zero/dynamic ice limit.  This change includes adding a new
mask bit for the cells that are on the Albany boundary and using that in
the interface.