iceThicknessHydro as mean of neighboring cells
iceThicknessHydro now smooths the entire ice thickness field used by the
subglacial hydrology model, instead of only smoothing along the
boundaries. In the subglacial hydrology model, each cell's surface elevation
is now a mean of its neighbors. As done previously, an additional
requirement prohibits the surface elevation used by the subglacial
hydrology model from being a local minimum or maximum. The purpose of this commit is to address
issues arising from single-cell subglacial lakes that form when the
surface elevation of a cell is lower than all of its neighbors,
particularly when coupling to ice dynamics.