Disable channels within subglacial lake interiors
This commit inhibits the development of channels within subglacial lake
interiors. Channels may not form on edges that connect two cells that
both have a waterThickness greater than config_SGH_bed_roughness_max.
Edges with only one adjacent lake cell are still able to form channels
so that the lake may drain.

Channels should not be forming on top of subglacial lakes, and we
hope that by inhibiting this unphysical situation that we can prevent
certain channel instabilities from occurring.