No waterPress. or  hydroPot. where no grounded ice
This commit zeros out waterPressure and and hydropotential where there
is no grounded ice. This is meant to resolve issues at the grounding
line occuring when the grounding line retreats to deeper waters so that
the direction of the hydropotential switches and points up glacier. With
this commit, we now have undefined waterPressure and hydropotential in
the ocean, which should negate this issue.