Small but positive minimum outflow at GL
Changes conditions at grounding line that disable inflow from the ocean.
In cases where inflow would occur, hydropotentialBaseSlopeNormal and
hydropotentialSlopeNormal are set to a negligibly small outflowing
value, instead of zero. This should prevent instabilities from forming
in the channel model that occur when the hydropotential slope is zero
and waterFlux and channelDischarge turn off.