Remove effectivePressure set to zero at timeinterval
Removes line that sets effectivePressure equal to zero at the start of
each timecycle; otherwise, effectivePressure was zero until the very end
of the timecycle. Instead, effectivePressure is now initialized to be equal
to effectivePressureSGH if there is no effectivePressure field in the input
file. A defualt value of -1e36 is given to effectivePressure to flag
whether the field is empty at the start of the run.