Only update halos when running out of halos
This will reduce the time spent on halo updates.  Need to loop over
nCells/nEdges instead of just *Solve to ensure the halos get updated
too.  This increases compute while reducing time spent on halo updates.
Because of vector reconstruction operations, 4 extra halos are needed,
which means to get any benefit from this, nHalos needs to be larger than
that, making it a hard tradeoff.

There may be an intermediate option where some of the halo updates are
performed every iteration and others are not, but that would require
more investigation.

For the hydro-radial test case on 3 procs, the baseline cost is:
                   time          num calls
subglacial hydro   0.03061         1
halo updates       0.01404        35

With these changes and nHalos=8, cost is:
subglacial hydro   0.04246         1
halo updates       0.00749        11

So, not cost effective.  Different behavior may occur for other
configurations (mesh, hydro state, num halos).