Remove sheet flow from pressure freeze calculation
I'm concerned there is a sign bug and combining the sheet and channel
terms together masks that.  Further, it's not obvious that the sheet
flow should be included in the pressure freeze calculation, as in
principle it's meant to be perpendicular to the channel, meaning it
would have a different derivative thatn used here.  The sheet term is
included in Werder et al (2013) but excluded from de Fleurian et al
(2018).